+++ CHANGES.md
@@ -1,5 +1,6 @@
 ### Skin Composer Version 46 ###
 * Added CrushMode options to TenPatch.
+* Added resizable preview to the "More Settings" dialog of TenPatch.
 
 ### Skin Composer Version 45 ###
 * Added commandline option -swingfd to use swing.JFileChooser for platforms that can't use Tiny File Dialogs.